+++ src/common/DayEventRenderer.js
@@ -31,6 +31,11 @@ function DayEventRenderer() {
 	var renderDayOverlay = t.renderDayOverlay;
 	var clearOverlays = t.clearOverlays;
 	var clearSelection = t.clearSelection;
+
+	if (t.getRowMaxWidth) {
+		var getRowMaxWidth = t.getRowMaxWidth;
+	}
+
 	var getHoverListener = t.getHoverListener;
 	var rangeToSegments = t.rangeToSegments;
 	var cellToDate = t.cellToDate;
@@ -203,8 +208,8 @@ function DayEventRenderer() {
 			var leftFunc = (isRTL ? segment.isEnd : segment.isStart) ? colContentLeft : colLeft;
 			var rightFunc = (isRTL ? segment.isStart : segment.isEnd) ? colContentRight : colRight;
 
-			var left = leftFunc(segment.leftCol);
-			var right = rightFunc(segment.rightCol);
+			var left = leftFunc(segment.leftCol, segment.gridOffset);
+			var right = rightFunc(segment.rightCol, segment.gridOffset);
 			segment.left = left;
 			segment.outerWidth = right - left;
 		}
@@ -616,6 +621,25 @@ function DayEventRenderer() {
 		});
 	}
 
+	function daySegSetLefts(segs, rowLefts) { // also triggers eventAfterRender
+		if (!rowLefts) {
+			return;
+		}
+
+		var i;
+		var segCnt = segs.length;
+		var seg;
+		var element;
+		var event;
+		for (i=0; i<segCnt; i++) {
+			seg = segs[i];
+			seg.left = rowLefts[seg.row] + (seg.left||0);
+			element = seg.element;
+			if (element) {
+				element[0].style.left = seg.left + 'px';
+			}
+		}
+	}
 	
 	function resizableDayEvent(event, element, segment) {
 		var isRTL = opt('isRTL');
@@ -648,24 +672,24 @@ function DayEventRenderer() {
 			var eventEnd;
 			var helpers;
 			var eventCopy = $.extend({}, event);
-			var minCellOffset = dayOffsetToCellOffset(dateToDayOffset(event.start));
+			var minCellOffset = t.dayOffsetToCellOffset(dateToDayOffset(event.start));
 			clearSelection();
 			$('body')
 				.css('cursor', direction + '-resize')
 				.one('mouseup', mouseup);
 			trigger('eventResizeStart', this, event, ev);
-			hoverListener.start(function(cell, origCell) {
+			hoverListener.start(function(cell, origCell, rowDelta, colDelta) {
 				if (cell) {
 
-					var origCellOffset = cellToCellOffset(origCell);
-					var cellOffset = cellToCellOffset(cell);
+					var origCellOffset = t.cellToCellOffset(origCell);
+					var cellOffset = t.cellToCellOffset(cell);
 
 					// don't let resizing move earlier than start date cell
 					cellOffset = Math.max(cellOffset, minCellOffset);
 
 					dayDelta =
-						cellOffsetToDayOffset(cellOffset) -
-						cellOffsetToDayOffset(origCellOffset);
+						t.cellOffsetToDayOffset(cellOffset) -
+						t.cellOffsetToDayOffset(origCellOffset);
 
 					eventEnd = getEventEnd(event).add('days', dayDelta); // assumed to already have a stripped time