+++ changelog.md
@@ -1,3 +1,15 @@
+v1.6.4-yearview 1.6 (2014-02-15)
+--------------------------------
+
+- Full support of a new 'year' view (draggable, selectable)
+- yearColumns option to customize layout, default is 3(x4)
+- firstMonth option for school year month offset
+- lastMonth option to show more than 12 months
+- hiddenMonths option to hide months in year view
+- new options: minDate/maxDate for all views (John Bafford)
+
+Implementation by tpruvot@github, based on Tim Martin idea.
+
 
 v2.0.0-beta2
 ------------